+++ flutter-hms-analytics/android/src/main/java/com/huawei/hms/flutter/analytics/viewmodel/HMSAnalyticsViewModel.java
@@ -37,7 +37,7 @@ import java.util.Set;
  */
 public class HMSAnalyticsViewModel implements HMSAnalyticsContract.Presenter {
 
-    //HiAnalytics instance
+    // HiAnalytics instance
     private final HiAnalyticsInstance analyticsInstance;
 
     public HMSAnalyticsViewModel(Context context) {